Add explicit "single-ESC" key for entering JDB
When an ESC sequence is sent to the kernel, the leading
ESC was interpreted as a request to enter JDB. However,
in such cases of sequences, this should not happen but
the sequence should be passed on. There's already
functionality to detect whether the ESC is coming alone
or in a sequence however we did not use that information
for entering JDB. This is done now by using an explicit
KEY_SINGLE_ESC value and checking for this in the
appropriate location.
With this, it is possible now to enable
CONFIG_FEATURE_EDITING_ASK_TERMINAL in busybox.

Change-Id: I98e8f4e8877e46573001da46a67c02838b175535