Fix logic for FPU handling in IPC
This fixes a bug that sneaked in originally when adding ARM HYP mode
support. And was further manifested when we added parenthesis at the
wrong place later to fix a warning. The logic for FPU transfer is that
the sender must set the transfer_fpu flag in the message tag and the
receiver must either set the inherit_fpu flag in its buffer descriptor
or must be waiting for a reply to some exception IPC.

Change-Id: I009c0c406296d46dafd081555b383b373234da99