fiasco/ia32: when configured for IBRS, panic on non-Intel CPUs
IBRS/IBPB are spectre mitigations for Intel CPUs. On Intel
CPUs, Fiasco does check if the CPU supports it, and panics if
support is missing.

This change changes the behaviour on non-Intel CPUs such that
if Fiasco is configured with IBRS/IBPB, it will panic and thus
give a meaningful message.

Previously, on non-Intel CPUs, Fiasco would just write to the
MSR which will generate a #GP and that caused Fiasco to reboot
without a meaningful message.

Change-Id: I4c3ed9c2d0dd6d04e4730b12dc14104f3ed653d0