fiasco/intel: deploy foreshadow mitigation if available
This change implements support for one mitigation for the
foreshadow Intel CPU bug.

It detects availability of the IA32_FLUSH_CMD MSR. This MSR
allows the OS to flush the L1D cache. This change implements
that this is done on each vmresume.

This does not help when SMT is used.

This will only work with current microcode.

Change-Id: I63b27ef8f0fb789435b23bc88a0276d576126f80