fixed Jdb::peek_task() for arm/mips/ppc32/sparc32
First, consider the size of the data type which is passed to the
function, in particular don't write an Mword in all cases. Second, the
special handling for data with sizeof >= 2 was not complete as accesses
crossing the Mword boundary were not considered. And last but not least
the access for 4 and 8 bytes worked only if the access is aligned. While
being at it we change the return type of access_mem_task() to 'unsigned
char *' for easier use of the result.

Note: Unaligned memory access crossing a page boundary still doesn't
work but usually this case isn't required.

Change-Id: Ia5ed5cc62be221d7e4fd1c069eaac49ab8964015