jdb_tcb: don't re-init stack view on whole screen
When changing to a different level, for example pressing ENTER to dump
the address referenced by the current stack position or pressing 'u' to
disassemble memory and coming back to the TCB view, don't re-initialize
the stack view. The changed cursor position of the highlighted stack
item was kind of annoying.

Change-Id: I55b5318f599cbbe54ce6a1c35422916955203798