ia32: implement ACPI reset
This change implements ACPI reset as a method to reboot the
machine. When a reboot is triggered from Jdb, this will be the
first method tried on amd64.

In L4Re, IO implements the ACPI reset itself, and will only
call into the kernel if it failed.

This change adds the ACPI reset method to system_shutdown(1)
as well, to ensure reboot works even when the userland does
not have proper ACPI support.

This fixes reboot on my test machine.

Change-Id: I65638a891259b7cd6c0a66a2b5fdee60cb8d1053