jdb: provide generic kdb_ke() for non-debug configurations
Even with JDB not configured the kernel may use kdb_ke() to enter the
kernel debugger. The behaviour of the kernel is inconsistent across
architectures. On x86 the kernel still issues an int3 which is not
desired in that case.

This change adds a generic implementation of kdb_ke* for the non-debug
case.  Instead a message is printed and the kernel resumes normal
operation.

Change-Id: Iba8da8d28e16ea9edaeabeb56b536fd17d1b3782