Atomically test and reset caller field in communication partner
When the receive phase of a call times out the calling thread resets
the caller field in the IPC partner. It is checked before whether the
calling thread is actually the one in the caller field.

If between the check and the reset another call arrives at the partner
the wrong caller is overwritten by zero, effectively removing the reply
capability the partner has for another thread.

This adds a method to test and reset the caller atomically only if it
is the expected value. This function is used in the do_ipc method to
prevent the aforementioned race.

Also move the reset_caller functions to the same place as the
set_caller function.

This fixes #1032

Change-Id: Ia8644fa65c869a71de913915cef8c62fed1750cc