Fix alignment in RPC protocols
GCC version 8 and clang at least since version 9 changed the definition
of C++ alignof() operator from preferred to minimum alignment. This
changes the value for some 64bit types on 32bit architectures.

This change switches to __alignof__ which according to the GCC manual
remains to return the preferred alignment of a type. Our user-level RPC
framework also uses __alignof__ so it is consistent with user-level and
it is also backward compatible when using different compiler versions
for user-level and kernel.

Change-Id: I3dd012ec01b37cf7cdd4417136dd343bb75d6b8c