Relax register pressure in asm statement with -O0
For -O0 builds with CONFIG_JDB and without CONFIG_NO_FRAME_PTR, the
compiler is unable to satisfy the asm statement constraints in
Thread::call_nested_trap_handler; presumably because it runs out of
available registers.

This commit alleviates the register pressure by moving two registers
from the clobers list explicitly to the output operand list, eliminating
thus the need for two extra registers. This fixes the -O0 build with the
above configuration.

Change-Id: I4ae1ec205bc1027c95ab0a7eabd633659e1545c2