perfctr: -Wmissing-declarations fix
In theory we should declare a separate header file for the function
prototypes and include that header file from the implementation in
the library and from the file where these functions are used.

OTOH these declarations belong to the library but we want to compile
the kernel even if the directory with the perfctr library is not
present.

Change-Id: If3ba4d902f9962ecfbba1629d71b350a91c27ff2