MIPS: fix potential delay of the timer interrupt
This happens mainly when running on QEMU. The old code already had
some detection mechanism but it did not consider the case when the
QEMU process got preempted between reading the counter and setting
the compare value. After setting the compare value we have to re-read
the counter to verify that the compare register is still beyond the
counter. If not, adapt the compare register and try again. Without
this check we would need to wait for a complete wrap around of the
counter to receive the next timer interrupt.

Fixes ticket #CD-102.

Change-Id: I29e28f91b4bb4f29642858df21406f750740fd09