Fix deadlock in Context::take_cpu_offline()
Context::take_cpu_offline() also drains the pending request queue. It
takes the queue's spinlock and calls handle_requests(), which must be
called without the queue lock held.

Re-organize take_cpu_offline() to not call handle_requests() with the
queue's lock held, while maintaining the need to hold the lock when
marking the CPU as offline.

Change-Id: I079228cb2717aa1d936a2a01074383a29ad0be2d
Signed-off-by: Christian Ludwig <cludwig@genua.de>