JDB: improve log entries for deleted threads
Currently JDB displays '????' as ID when it renders the trace buffer.
This change prints the original TCB address prefixed with 'p:' in case
there is no debug ID found for the TCB. The 'p:' is to disambiguate
between debug IDs and TCB addresses.

NOTE: There might be invalid debug IDs printed for thread that does
not exist any more, but a new object is allocated at the same address.

Change-Id: Ib1b2ed81d462aed8c9c2c911434f7d96066c21a8