vgic_v3: use neutral macro name
The macro should have the same name for reading and writing the LRs
because the lines should look identical for Gic_h_v3::save_lrs() and
Gic_h_v3::load_lrs(), despite the fact that we read the LRs in the
first case and write the LRs in the second case.

Using a TRANSFER_LRS macro calling a dedicated TRANSFER_LR macro for
both functions is not (easily) possible due to preprocess.

Change-Id: I07addb1e868a6de7276753ab31e91f243ba4a00d