ARM, MIPS: fix eager FPU usage
On x86 eager FPU switching is the default. For that to work the area to
store the FPU state has to be allocated separately from
Fpu_alloc::alloc_state(). This allocation is missing for ARM and MIPS.
As a result an in-kernel assertion is triggered if an eager FPU
configuration is used on these architectures.

This change adds the allocation of the FPU state buffer and enables
the usage of eager FPU switching on ARM and MIPS as well.

Change-Id: Ic0fd11ff2e614786ccae23b5ea60a5e461ce8165