uart_16550: re-add special cases for X86 and PXA
As of Change-Id:Ic6072d9d6f2c13b5f11f034f271958b371fb1664 F_skip_init
is no longer used, thus remove it.

During UART initialization we need to set bit 3 ('Auxiliary output 2')
in the MCR register on X86. On PXA we need to set bit 6 of the IER
register. These special cases are already considered by bootstrap and
so far the F_skip_init flag prevented the re-initialization of the MCR
register and the IER register in the kernel for these platforms.

However, this approach did never work during the re-initialization of
the UART during resume (and that's why F_skip_init was removed). In a
future approach bootstrap will pass the register initialization values
to the kernel which will be then able to use them during resume.

For now just pass dedicated initialization values to the Uart_16550
constructor for these two platforms.

Change-Id: I29f5fc64bfd6f43d504713c3e8dbeba5cf0e5b2d