ARM: Use generic timer as time stamp counter
So far PMCCNTR_EL0 is used as time stamp counter for the tracebuffer.
The ARM generic timer has two benefits: It has a known frequency and
it seems to work better on QEMU.

As we now have a known frequency we can also display proper times in
the tracebuffer view.

Change-Id: Id9d8a08d0a748d8788358a08478efff142931e7a