ARM: increase maximum tracebuffer size
So far the maximum number of tracebuffer events for 64-bit hosts was
16,384. With a size of 128 bytes per event this requires 2MB.

The new limit is 32MB which is sufficient for 262,144 events. Because
tracebuffer memory reduces the amount of available kernel memory, the
initial size of the kernel memory pool needs to be increased as well.
For ARM this is Config::KMEM_SIZE with a current value of 16MB.

Use the -tbuf_entries=<number of events> Fiasco command line parameter
to increase the default value.

Increasing this limit works only for ARM64 with virtualization enabled.

Change-Id: I363e333f071c684ad5353ce3534f78f5c91a0578