AARCH64: Do not trap debug exceptions ein EL0/EL1
Change I4b56c1acc520d3bd07af05a1d0bab613eb777e4b changed the
initialization of MDCR_EL2 to include the TDE bit. If this bit
is set to 1
* debug exceptions generated at EL1/EL0 are routed to EL2
* All accesses to debug registers ... are trapped to EL2

Since UVMM currently does not handle break point exceptions it
aborts when a guest executes a brk instruction, which would
be routed to the UVMM as break point exception.

Here we remove the TDE bit from the initialization for the time.

Change-Id: Id3e530230f5cece686dbf68e885b397ff988eb6c