Fix factory issues
Fix possible overflow when adding the requested limit to the size of the
factory object itself.

Handle not-yet-empty factory deletion correctly.

This change also prevents allocation of new unlimited factories, which
was never intended. And also prevents allocation of new kernel memory
on a factory that is already in deletion.

Fixes tickets #CD-305, #CD-306, #CD-307.

Requires-Change-Id: I78bd4f8671248310f30a5bd708f1dca4a3abf206
Change-Id: I2be9a74c55a89fd2ba4135d13b67d9ff5950e901