L4_error::error() also consider the LSB
The least significant bit of the error code is used to distinguish
between error codes related to the IPC send phase (LSB=0) and the
receive phase (LSB=1).

To be able to test for error codes related to the IPC receive phase,
consider the least significant bit as well.

At the moment this makes only a difference in

  Thread_object::sys_vcpu_resume()

where the error code is explicitly checked against R_timeout (0x03).

Fixes #CD-310.

Change-Id: I36d9dee06d48f960d20f690c9206caa177bc3708