MIPS: fix clobbers in switch_cpu
We have to clobber really all GP registers except s8/fp which is
explicitly saved/restored as part of the Assembler code.

Without this change the compiler could assume that $6/a0 and $7/a1
are unchanged. We now explicitly use 4 registers as input operands
and mark all of them as modified.

Also rename '_current' => '_old_this' and '_to' => '_new_this' to
make the naming more obvious (and do it similar to ARM).

Change-Id: Ib1e535286f758f84511386fdbd3c8dcc14b51194