fix off-by-one in Intel IOMMU address width calculation
The maximum DMA physical addressability supported by an Intel IOMMU is
indicated via the host address width field in the DMA remapping
reporting structure. The reported value is N but the supported host
address width has to be calculated as (N + 1) (see Intel IOMMU spec,
section 8.1). Missing that results in making the upper half of the DMA
space unaccessible and causing mappings to fail.

This change corrects the calculation of the host address width and makes
the whole DMA space usable.

Thanks to Christian Ehrhardt from Genua for reporting this issue.

Change-Id: Ibb05b5ad8491cc99f6de4cb6493988a1b34abb1c