ARM: make sure GICv3 redistributor is awake
The reset state of the redistributor is a low power state. After power
up, the system software must wake the redistributor to forward
interrupts to the CPU. Usually this is done already by the ARM trusted
firmware but on systems without any firmware the redistributor will
still be asleep when Fiasco boots.

Change-Id: Iac32164555d433eada9c68fe734889792f27bd22
Requires-Change-Id: I354e3284cfc67a6696fdf614737a7c7e5e352264