fiasco/x86: Do not forward hardware NMIs to vCPUs
When a hardware NMI occurs and a vCPU thread happens to execute in one
of its userland tasks, do not reflect that random event to the vCPU's
exception handler.

Always forward NMI events to JDB. If JDB was not compiled in, this will
log a message and resume execution.

This also fixes inconsistent behavior, NMIs are not reflected when the
vCPU executes in its kernel task.

Change-Id: Ic97c9a2179771d1e70a2bc28861e2da162108211