Add unittest kconfig and build system
This change adds a new subsystem for unit tests called 'utest' and a
kconfig option to enable the unit tests.

The unit tests hook to the weak symbol `init_unittest()` of
kernel_thread.cpp. If the function is defined, it's invoked instead of
`init_workload()`.

All unit tests shall be placed in a subdirectory of test/utest/. Each
subdirectory must contain a separate Modules.utest file extending the
INTERFACES_UTEST variable with the tests in its directory.
See the description in the Modules.utest file for details.

The preprocess and compile steps generate all object files for each
unit test.
In the link step the script generates new targets which include the
.o-files of the unit test, thus defining `init_unittest()` for the
resulting binary.

The name of the CHECKSUM target is changed, as it conflicts with
targets generated from checksum.cpp.

Change-Id: Idc3631a9e1eccecba7c76ae350fe3a44cd7d9697