MDB: refactoring...
Add a MDB::Frame::clear() function to actually clear the lock
of the frame and mark the frame as invalid. (For Mapdb this is
actually a rename from rest() to clear()).
Also add MDB::Frame::might_clear() interface that clear a frame
only if it is valid.
Add MDB::Frame::same_lock() inertafce to check if two frame actually
use the same lock (this is used by the overmap logic).

Change-Id: Ia9f856a45b25ac87565869790a0d75d09d3c21c0