arm: do not use condition on potentially undefined insn
It is partially implementation defined behaviour if an instruction that
fails the condition check executes as NOP or raises an undefined
instruction exception. Specifically, if the instruction itself is
undefined and the exception cause does not depend on the data values,
an implementation is permitted to raise an undefined instruction
exception. See Armv8-A architecture reference manual section G1.16.1 -
"Conditional execution of undefined instructions".

Change-Id: Ia4301c79b27554fcf5c967b5e62a9a3c65a80797