utest_fw: print a single TAP line per test
We print just a single TAP line per test. Multiple failure messages in
a single test only print their failure comments, but not a TAP line.
For each TAP line, the printed number value after ok/not ok is the
number in the execution order.

new_test: At the beginning of a test, the writer needs to call
          new_test() giving a group and a test name. This marks
          the end of the previous test. If the test has the same
          group and test name as the previous test, a instance
          counter is incremented to uniquely identify each single
          test.

tap_todo: A test can be marked as TODO using the tap_todo() function.
          tap_todo must be called after new_test() but before the
          first UTEST_ macro!

The message given to each assert/expect statement is now printed as a
comment and only in case the statement fails.

If debug verbosity is enabled (see Utest_debug) each new test will print
it's name and succeeding UTEST_ macros print their info and location.

Change-Id: I768354d39235674e51db404ce9ccb26bbf30787a