Optimize memcpy/memset(_mwords)
These standard functions are called in performance-critical sections
and are adding significant overhead because the default implementation
copied/set only byte-by-byte.

The first part of the optimization is to optimize the default
C-implementations so that they yield decent assembly on most platforms.
Anything should be better than copying bytes only. For the
implementation we can also exploit the fact that unaligned accesses are
practically non-existent in Fiasco and can spare all the special
handling. This affects all platforms except memcpy on IA32 which has a
dedicated inline assembly version already.

For the arm 32-bit platform the memcpy_mwords function is additionally
optimized because the compiler is not smart enough to use ldm/stm by
itself. As almost all call sites use a constant length, the code is
especially optimized for this case.

The memset_mwords function on arm stores 64 bit words to nudge the
compiler to exploit the strd instructions. Even though not strictly
required by the architecture, aligning the pointer to 64 bit will
generate more efficient bus access on write-through caches. On arm64
memset_mwords stores multiple 64 bit words at once to make sure the
compiler uses stp where possible.

Change-Id: I583c3d6c103f1110f1f34787d5fb05ce7e168bbc