ARM: link bootstrap as ET_REL not ET_EXEC
Support for linking of ET_EXEC section was deprecated in binutils 2.35
and dropped in binutils 2.36.

Instead, extract the raw section of the fully linked bootstrap object
and create a relocatable object which can be linked to the kernel. The
required symbols for the linker are extracted separately.

Fixes ticket #CD-301.

Change-Id: I0cdc2aacb5dbd01677d93e2bb1103940ac60e848