Scheduler::sys_run(): fix UTCB read after check
The previous code copied parts of the UTCB after some checks. However,
between checking and copying, the UTCB might have changed. This commit
fixes this issue by first copying and then checking the copy.

The parameters are now checked early in 'Scheduler::sys_run()' because
it is not possible to propagate error codes from 'Sched_context::set()'
back to userland. Also change the return type of 'Sched_context::set()'
to reflect that the return code of these functions is not evaluated by
the callers.

Fixes #CD-420.

Change-Id: I7ae5ac6051d24d912513da5b520c1e9bd8d0eff0