utest: introduce function to SKIP tests
The TAP format allows to record a test as skipped. This means the test
was not executed, therefore produced no failures and thus is sucessful.

This commit introduces Utest_fw::tap_skip() to record this state for the
current test.
This function must be called before any UTEST_ macro is executed.
Otherwise, the test result is not guaranteed to be recorded
as OK-SKIP.
It's on the caller to terminate the test after the call to tap_skip().

This also fixes stale documentation of tap_todo().

fixes ticket #TEST-169.

Change-Id: I22cf24a6c64c671f2f8953bf67c5d423b0b4ad3b