div32/mod32: use 'static inline' not 'extern inline'
Mark these functions 'static inline' instead of 'extern inline'.

- 'extern inline':
  A function marked 'extern inline' will always appear as a single
  separate function in the final binary, even if all instances were
  successfully inlined. Even with CONFIG_INLINE not set, these small
  functions are _usually_ inlined.

- 'static inline':
  If the function can always be inlined, no separate implementation
  is generated for the final binary. If inlining is not possible,
  the compiler generates one implementation per file.

- 'inline':
  With C++, the compiler behaves similar as if used 'static inline'.
  However, if inlining is not possible with C code, the compiler does
  NOT generate any function implementation but relies on an external
  function implementation.

Unfortunately we cannot use 'inline' for div32()/mod32() because these
functions are also called from C code.

If a function can be usually inlined, 'static inline' is preferred
for C code and 'inline' for C++ code. If a function is normally too
big for inlining, 'extern inline' is preferred for C code (to not
have multiple copies of that function) and 'inline' is preferred for
C++ code.

Change-Id: Ie8f3318d444ccfcab5ec1ab474bb5809974db575