mapdb: Rename parameters to better reflect use
The page_shifts_max parameter is used for the Mapdb to indicate how many
elements the page_shifts array contains. The suffix _max however
suggests that this would be the maximum array index, which is not the
case. Rename to _num to better reflect the use of the number.

The same is also true for Treemap's sub_shifts_num and shifts_num
parameters and member variables.

Change-Id: I14857d3a78878e26465e6e8f59278f0aa1985285